Create Index.html.erb
Created file includes some functional embedded Ruby which allows the todo list to populate as todo items are created.  

It also includes checkboxes to be "checked" when a todo item is completed.

It also allows todo items to be removed or destroyed when the x on the right is clicked.  This removes it from database too.